Changes in usage 1. Rename App to EditorApp 2. Add option to provide width and height as prop 3. Add option to provide canvasId as prop 4. EditorApp now uses only context. 5. EditorApp exposes custom renderer props for Plugin, Action and property.
